+++ NEWS.md
@@ -1,6 +1,7 @@
 # SimpleUpset 0.0.4
 
 * Changed set & intersect sorting to use data masking instead
+* Switched size columns to be 'size' instead of 'n'
 
 # SimpleUpset 0.0.3